refactor: break down room_list.dart into smaller focused files
Extract 7 files from the 1300-line room_list.dart monolith:
- room_list_models.dart: ListItem sealed class hierarchy
- room_list_builder.dart: section-building logic and search helpers
- room_tile.dart: RoomTile widget
- invite_tile.dart: InviteTile widget
- message_search_tiles.dart: MessageSearchHeader, MessageSearchResultTile, LoadMoreButton
- speed_dial_item.dart: SpeedDialItem widget
- room_section_header.dart: RoomSectionHeader widget

room_list.dart retains RoomList StatefulWidget with search/FAB animation
state and the build method (~280 lines).
