+++ lib/widgets/room_list_builder.dart
@@ -1,0 +1,174 @@
+import 'package:matrix/matrix.dart';
+
+import '../models/space_node.dart';
+import '../services/matrix_service.dart';
+import '../services/preferences_service.dart';
+import 'room_list_models.dart';
+
+// ── Section-building helpers for the room list ──────────
+
+bool roomMatchesQuery(Room r, String q) {
+  if (r.getLocalizedDisplayname().toLowerCase().contains(q)) return true;
+
+  final alias = r.canonicalAlias;
+  if (alias.isNotEmpty && alias.toLowerCase().contains(q)) return true;
+
+  final dmPartner = r.directChatMatrixID;
+  if (dmPartner != null && dmPartner.toLowerCase().contains(q)) return true;
+
+  return false;
+}
+
+List<Room> applySearch(List<Room> rooms, String query) {
+  if (query.isEmpty) return rooms;
+  final q = query.toLowerCase();
+  return rooms.where((r) => roomMatchesQuery(r, q)).toList();
+}
+
+Set<String>? spaceRoomIds(MatrixService matrix) {
+  final selectedIds = matrix.selectedSpaceIds;
+  if (selectedIds.isEmpty) return null;
+
+  final ids = <String>{};
+  void collect(SpaceNode node) {
+    ids.addAll(node.directChildRoomIds);
+    for (final sub in node.subspaces) {
+      collect(sub);
+    }
+  }
+  for (final node in matrix.spaceTree) {
+    if (selectedIds.contains(node.room.id)) collect(node);
+  }
+  return ids;
+}
+
+List<ListItem> buildSectionItems(
+  MatrixService matrix,
+  PreferencesService prefs,
+  String query,
+) {
+  final collapsed = prefs.collapsedSpaceSections;
+  final selectedIds = matrix.selectedSpaceIds;
+  final tree = matrix.spaceTree;
+  final items = <ListItem>[];
+
+  // Invited rooms at the top (filtered by search)
+  final invitedRooms = applySearch(matrix.invitedRooms, query);
+  for (final room in invitedRooms) {
+    items.add(InviteItem(room: room));
+  }
+
+  final pinnedIds = <String>{};
+
+  if (selectedIds.isNotEmpty) {
+    // Space selected: show only that space's rooms with subspace hierarchy
+    final visibleNodes = tree
+        .where((n) => selectedIds.contains(n.room.id))
+        .toList();
+    for (final node in visibleNodes) {
+      _addSpaceSection(items, node, 0, matrix, collapsed, pinnedIds, query);
+    }
+  } else {
+    // No space selected (Home): Pinned → DMs → Unsorted
+
+    // Pinned section
+    final pinnedRooms = applySearch(
+        matrix.rooms.where((r) => r.isFavourite).toList(), query);
+    pinnedIds.addAll(pinnedRooms.map((r) => r.id));
+    if (pinnedRooms.isNotEmpty) {
+      items.add(HeaderItem(
+        name: 'Pinned',
+        sectionKey: PreferencesService.pinnedSectionKey,
+        depth: 0,
+        roomCount: pinnedRooms.length,
+      ));
+      if (!collapsed.contains(PreferencesService.pinnedSectionKey)) {
+        for (final room in pinnedRooms) {
+          items.add(RoomItem(room: room, depth: 0));
+        }
+      }
+    }
+
+    // DMs section — all direct chats
+    final dmRooms = applySearch(
+        matrix.rooms.where((r) => r.isDirectChat && !pinnedIds.contains(r.id)).toList(), query);
+    if (dmRooms.isNotEmpty) {
+      items.add(HeaderItem(
+        name: 'Direct Messages',
+        sectionKey: PreferencesService.dmSectionKey,
+        depth: 0,
+        roomCount: dmRooms.length,
+      ));
+      if (!collapsed.contains(PreferencesService.dmSectionKey)) {
+        for (final room in dmRooms) {
+          items.add(RoomItem(room: room, depth: 0));
+        }
+      }
+    }
+
+    // Unsorted section (orphan group rooms)
+    final orphans = applySearch(matrix.orphanRooms, query)
+        .where((r) => !pinnedIds.contains(r.id) && !r.isDirectChat)
+        .toList();
+    if (orphans.isNotEmpty) {
+      items.add(HeaderItem(
+        name: 'Rooms',
+        sectionKey: PreferencesService.unsortedSectionKey,
+        depth: 0,
+        roomCount: orphans.length,
+      ));
+      if (!collapsed.contains(PreferencesService.unsortedSectionKey)) {
+        for (final room in orphans) {
+          items.add(RoomItem(room: room, depth: 0));
+        }
+      }
+    }
+  }
+
+  return items;
+}
+
+void _addSpaceSection(
+  List<ListItem> items,
+  SpaceNode node,
+  int depth,
+  MatrixService matrix,
+  Set<String> collapsed,
+  Set<String> pinnedIds,
+  String query,
+) {
+  final rooms = applySearch(matrix.roomsForSpace(node.room.id), query)
+      .where((r) => !pinnedIds.contains(r.id)).toList();
+
+  // Count total rooms including all nested subspaces for the header
+  var totalRooms = rooms.length;
+  void countSubspaces(List<SpaceNode> subs) {
+    for (final sub in subs) {
+      totalRooms += applySearch(
+          matrix.roomsForSpace(sub.room.id), query)
+          .where((r) => !pinnedIds.contains(r.id)).length;
+      countSubspaces(sub.subspaces);
+    }
+  }
+  countSubspaces(node.subspaces);
+
+  // Skip entirely empty sections
+  if (totalRooms == 0) return;
+
+  items.add(HeaderItem(
+    name: node.room.getLocalizedDisplayname(),
+    sectionKey: node.room.id,
+    depth: depth,
+    roomCount: totalRooms,
+  ));
+
+  if (!collapsed.contains(node.room.id)) {
+    for (final room in rooms) {
+      items.add(RoomItem(room: room, depth: depth));
+    }
+    for (final sub in node.subspaces) {
+      _addSpaceSection(
+          items, sub, depth + 1, matrix, collapsed, pinnedIds, query);
+    }
+  }
+}
